+++ CHANGELOG.md
@@ -24,6 +24,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Add `TestsFilePathTooLong` to similate long path inside test.
 - Update to use `ricaun.Revit.UI.Tasks` library to run async tests.
 - Update `ApplicationUtils` check local file exists.
+- Update `RevitTestConsole` with `SetCurrentDirectory` using the source test file.
 
 ## [1.3.4] / 2024-05-27
 ### Console